Conflicts:
	README.AHBOT
	addition/00019_mangos_creature_onkill_reputation.sql
	addition/00146_mangos_creature_spell_copy.sql
	addition/00148_mangos_creature_spell.sql
	addition/00234_mangos_spell_proc_event.sql
	addition/characters/99999_07_characters_addition.sql
	sql_mr/custom_characters_tables.sql
	sql_mr/custom_realmd_tables.sql
	sql_mr/custom_rerun_every_mangos_DB_update.sql
	sql_mr/innormative_words.txt
	sql_mr/mr00019_mangos_creature_onkill_reputation.sql
	sql_mr/mr00146_mangos_creature_spell_copy.sql
	sql_mr/mr00234_mangos_spell_proc_event.sql
	sql_mr/mr00242_mangos_creature_onkill_reputation.sql
	sql_mr/mr00393_mangos_spell_proc_event.sql
	src/game/AuctionHouseBot/AuctionHouseBot.cpp
	src/game/AuctionHouseBot/AuctionHouseBot.h
	src/game/AuctionHouseBot/ahbot.conf.dist.in
	src/game/AuctionHouseHandler.cpp
	src/game/AuctionHouseMgr.cpp
	src/game/AuctionHouseMgr.h
	src/game/CMakeLists.txt
	src/game/Chat.cpp
	src/game/ConfusedMovementGenerator.cpp
	src/game/ConfusedMovementGenerator.h
	src/game/FleeingMovementGenerator.cpp
	src/game/FleeingMovementGenerator.h
	src/game/GameObject.h
	src/game/GridNotifiers.cpp
	src/game/HomeMovementGenerator.cpp
	src/game/HomeMovementGenerator.h
	src/game/Language.h
	src/game/Level3.cpp
	src/game/Mail.cpp
	src/game/MovementGenerator.h
	src/game/Object.cpp
	src/game/Player.cpp
	src/game/PointMovementGenerator.cpp
	src/game/PointMovementGenerator.h
	src/game/SpellEffects.cpp
	src/game/TargetedMovementGenerator.cpp
	src/game/TargetedMovementGenerator.h
	src/game/Traveller.h
	src/game/Unit.cpp
	src/game/Unit.h
	src/game/WaypointMovementGenerator.cpp
	src/game/WaypointMovementGenerator.h
	src/game/World.cpp
	src/game/World.h
	src/game/playerbot/PlayerbotAI.cpp
	src/mangosd/mangosd.conf.dist.in
	src/shared/Log.cpp
	src/shared/Log.h
	src/shared/SystemConfig.h.in
	win/VC100/game.vcxproj.filters
	win/VC80/game.vcproj
	win/VC90/game.vcproj
	win/mangosdVC90.sln

+++ src/game/MapPersistentStateMgr.cpp
@@ -457,7 +457,7 @@ void DungeonResetScheduler::LoadResetTimes()
             Difficulty difficulty = Difficulty(fields[1].GetUInt32());
             uint64 _oldresettime = fields[2].GetUInt64();
 
-            if (_oldresettime > (time(NULL) + INSTANCE_MAX_RESET_OFFSET))
+            if (_oldresettime > uint64(time(NULL) + INSTANCE_MAX_RESET_OFFSET))
             {
                 MapDifficultyEntry const* mapDiff = GetMapDifficultyData(mapid,Difficulty(difficulty));
                 oldresettime = DungeonResetScheduler::CalculateNextResetTime(mapDiff, time(NULL));
